chore: Update install.sh script to retrieve domain IP address
This commit modifies the install.sh script to retrieve the IP address of the specified domain. It uses the `dig` command to fetch the IP address and then filters the result to extract the first IP address found. This change ensures that the correct IP address is obtained for further processing.